Changed the way in which circles were generated to be more efficient and faster.
By changing the over to python generators the speed up was considerable for each circle group. The only bottleneck now is from tkinter's canvas widget.

Plan on changing over to a custom GUI for future projects so that each pixel is not an object reducing overall speed of the program after a few hundred thousand objects are generated.